Hoist mode transition to ModeManager
In several cases, transitioning to a new mode requires a sum of
information that is available within the mode we are entering as well as
information available at the "global" ModeManager. In this commit, we
call mode transition logic from the ModeManager. To allow this, we add a
new mode state RequestingModeTransition, which serves as a message to
signal to the ModeManager whose initiate_mode_transition to call.

Caveat:
- To avoid huge amounts of boilerplate, we had to put each mode's
  handler behind an Arc. This is probably ok, and probably will save us
  from future headaches, as mode transition is probably going to get
  more complicated, not simpler. However, since there is so little
  processing in Arpad so far, it is not inconceivable that locking the
  Mutex will be a meaningful part of any overhead -- especially since we
  now need to lock for each and every message. It is possible we will
  need to avoid any lock that can be safely avoided, in which case we
  may need to forego the closure and write a lot more boilerplate.